move validate broker fn
