+++ crates/mqtt/src/broker_validator.rs
@@ -1,6 +1,7 @@
-use std::time::{Duration, Instant};
-
-use crate::util;
+use std::{
+    net::{Ipv6Addr, TcpStream, ToSocketAddrs},
+    time::{Duration, Instant},
+};
 
 #[derive(Default)]
 pub(crate) struct BrokerValidator {
@@ -43,7 +44,7 @@ impl BrokerValidator {
                 if let Err(e) = std::thread::Builder::new()
                     .name("broker-validator".into())
                     .spawn(move || {
-                        let result = util::validate_broker(&cp_host, &cp_port);
+                        let result = validate_broker(&cp_host, &cp_port);
                         if let Err(e) = tx.send(result) {
                             log::error!("{e}");
                         }
@@ -64,4 +65,38 @@ impl BrokerValidator {
             }
         }
     }
+}
+
+fn validate_broker(host: &str, port: &str) -> Result<(), String> {
+    // Validate port first
+    let port: u16 = port.parse().map_err(|e| format!("Invalid port: {e}"))?;
+
+    // Format host properly for IPv6 if needed
+    let formatted_host = if let Ok(ipv6) = host.parse::<Ipv6Addr>() {
+        format!("[{ipv6}]")
+    } else {
+        host.to_owned()
+    };
+
+    // Create proper address string
+    let addr_str = format!("{formatted_host}:{port}");
+
+    // Resolve hostname using DNS (including mDNS if supported by system)
+    let addrs = addr_str
+        .to_socket_addrs()
+        .map_err(|e| format!("DNS resolution failed: {e}"))?;
+
+    // Try all resolved addresses with timeout
+    let mut last_error = None;
+    for addr in addrs {
+        match TcpStream::connect_timeout(&addr, Duration::from_secs(2)) {
+            Ok(_) => return Ok(()),
+            Err(e) => last_error = Some(e),
+        }
+    }
+
+    Err(last_error.map_or_else(
+        || "No addresses found".to_owned(),
+        |e| format!("Connection failed: {e}"),
+    ))
 }
